Updated mel commands
Get Attr Shader: List all the attribute of the shader and shading group
Create Texture File: Create the nodes of a texture file with the path
Create 2D Placement: Create node for the 2d placements and connects to the texture.
Connect Atributes: Do the connections between two attributes
Create Bump: Create the bump node and connects the texture file
Create Displacement: Create the displacement shader and connect to the texture file
Dialog Window: Open a file browser from maya to select the textures.