Update mel_helper.py
get_all_shaders: added a list so it will not return all the shaders found.
create_bump: looks for the shader type to assign the correct node type.
create_displacement: added a new node for the connection between the texture and the displacement node.
Added new function to create a color correct, range and create different nodes as utility if needed.