Auto fill textures
Added a new feature to auto complete the textures following a pattern, to make the selection more easier for the user.
Shader Creator UI: Added a new function to clean the UI after creating a shader, reseting the auto search and added the function to auto fill the UI
Shader Creator: Updated the names in the UI to remove conflict.
Btn Actions: changes name of sssColor to remove conflict
Path Helper: Added the function to look in the same folder for the similar files of the file selected.